+++ resources/js/components/openaccess/App.vue
@@ -14,22 +14,22 @@
                     >
                     {{ state.projectName }}
                 </router-link>
-            </div>
-            <div
-                id="navbar-shortcuts"
-                class="collapse navbar-collapse"
-            >
-                <!-- Left Side Of Navbar -->
-                <ul class="navbar-nav me-auto">
-                    <li class="nav-item">
-                        <router-link
-                            :to="{ name: 'modules' }"
-                            class="nav-link"
-                        >
-                            Modules
-                        </router-link>
-                    </li>
-                </ul>
+                <div
+                    id="navbar-shortcuts"
+                    class="collapse navbar-collapse"
+                >
+                    <!-- Left Side Of Navbar -->
+                    <ul class="navbar-nav">
+                        <li class="nav-item">
+                            <router-link
+                                :to="{ name: 'modules' }"
+                                class="nav-link"
+                            >
+                                Modules
+                            </router-link>
+                        </li>
+                    </ul>
+                </div>
             </div>
         </nav>
         <div class="container my-3 col overflow-hidden d-flex flex-column">